move database setup to generic conftest.py
